+++ src/main/java/com/kobe/warehouse/domain/LotSold.java
@@ -2,4 +2,4 @@ package com.kobe.warehouse.domain;
 
 import java.io.Serializable;
 
-public record LotSold(long id, String numLot, int quantity) implements Serializable {}
+public record LotSold(int id, String numLot, int quantity) implements Serializable {}
