+++ src/main/java/com/kobe/warehouse/domain/ReponseRetourBonItem.java
@@ -22,7 +22,7 @@ public class ReponseRetourBonItem implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Column(name = "date_mtv", nullable = false)
@@ -47,11 +47,11 @@ public class ReponseRetourBonItem implements Serializable {
     @Column(name = "after_stock", length = 8)
     private Integer afterStock;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public ReponseRetourBonItem setId(Long id) {
+    public ReponseRetourBonItem setId(Integer id) {
         this.id = id;
         return this;
     }
