+++ src/main/java/com/kobe/warehouse/repository/AjustService.java
@@ -11,7 +11,7 @@ import org.springframework.data.domain.Pageable;
 public interface AjustService {
     Page<AjustDTO> loadAll(AjustementFilterRecord ajustementFilterRecord, Pageable pageable);
 
-    Resource exportToPdf(Long id) throws IOException;
+    Resource exportToPdf(Integer id) throws IOException;
 
-    Optional<AjustDTO> getOneById(Long id);
+    Optional<AjustDTO> getOneById(Integer id);
 }
