+++ src/main/java/com/kobe/warehouse/repository/CategorieRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface CategorieRepository extends JpaRepository<Categorie, Long> {
+public interface CategorieRepository extends JpaRepository<Categorie, Integer> {
     Optional<Categorie> findOneByLibelle(String libelle);
 }
