+++ src/main/java/com/kobe/warehouse/repository/DciRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DciRepository extends JpaRepository<Dci, Long> {
+public interface DciRepository extends JpaRepository<Dci, Integer> {
     Page<Dci> findAllByCodeContainingIgnoreCaseOrLibelleContainingIgnoreCaseOrderByLibelleAsc(
         String code,
         String libelle,
