+++ src/main/java/com/kobe/warehouse/repository/FamilleProduitRepository.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Repository;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface FamilleProduitRepository extends JpaRepository<FamilleProduit, Long>, JpaSpecificationExecutor<FamilleProduit> {
+public interface FamilleProduitRepository extends JpaRepository<FamilleProduit, Integer>, JpaSpecificationExecutor<FamilleProduit> {
     Page<FamilleProduit> findAllByCodeOrLibelleContainingAllIgnoreCase(String code, String libelle, Pageable pageable);
 
     Optional<FamilleProduit> findFirstByLibelleEquals(String libelle);
